+++ CHANGES.md
@@ -49,7 +49,6 @@ Changes in 10.0
 - SQL use DAYOFWEEK() for MySQL or cast(extract(DOW)+1 AS int) for PostrgeSQL, program wide
 - SQL cast(AS UNSIGNED) for MySQL or cast(AS INT) for PostgreSQL, program wide
 - SQL cast custom_fields ID as VARCHAR(10) instead of TEXT for MySQL compatibility in GetStaffList.fnc.php, GetStuList.fnc.php & Search.fnc.php
-- SQL cast CURRENT_DATE+X as DATE for MySQL compatibility in Reminders.php
 - SQL rename $field COLUMN (reserved keyword) to COLUMN_NAME for MySQL compatibility in CustomFields.fnc.php, GetStaffList.fnc.php, GetStuList.fnc.php & Search.fnc.php
 - SQL remove use of nextval in rosariosis_fr.sql
 - Rename $pg_dumpPath configuration variable to $DatabaseDumpPath in config.inc.sample.php & DatabaseBackup.php
@@ -64,6 +63,7 @@ Changes in 10.0
 - Rollback Fix PostgreSQL error invalid ORDER BY, only result column names can be used, program wide
 - HTML use number input for Gradebook config options in Configuration.php
 - SQL limit results to current school year in AddDrop.php
+- SQL always use INTERVAL to add/subtract days to date for MySQL compatibility in Reminders.php, Transactions.php, ServeMenus.php, Assignments.php, StudentGrades.php, Rollover.php & Portal.php
 
 Changes in 9.3
 --------------